+++ src/Microsoft.DotNet.XHarness.Apple/AppUninstaller.cs
@@ -9,7 +9,7 @@ using Microsoft.DotNet.XHarness.Common.Execution;
 using Microsoft.DotNet.XHarness.Common.Logging;
 using Microsoft.DotNet.XHarness.iOS.Shared.Execution;
 
-namespace Microsoft.DotNet.XHarness.iOS
+namespace Microsoft.DotNet.XHarness.Apple
 {
     public class AppUninstaller
     {
@@ -28,7 +28,7 @@ namespace Microsoft.DotNet.XHarness.iOS
         {
             var args = new MlaunchArguments();
 
-            for (int i = -1; i < _verbosity; i++)
+            for (var i = -1; i < _verbosity; i++)
             {
                 args.Add(new VerbosityArgument());
             }
