+++ src/Microsoft.DotNet.XHarness.Apple/DeviceLogCapturerFactory.cs
@@ -7,7 +7,7 @@ using Microsoft.DotNet.XHarness.Common.Logging;
 using Microsoft.DotNet.XHarness.iOS.Shared.Execution;
 using Microsoft.DotNet.XHarness.iOS.Shared.Logging;
 
-namespace Microsoft.DotNet.XHarness.iOS
+namespace Microsoft.DotNet.XHarness.Apple
 {
     public interface IDeviceLogCapturerFactory
     {
