+++ tests/Microsoft.DotNet.XHarness.Apple.Tests/AppInstallerTests.cs
@@ -16,7 +16,7 @@ using Microsoft.DotNet.XHarness.iOS.Shared.Hardware;
 using Moq;
 using Xunit;
 
-namespace Microsoft.DotNet.XHarness.iOS.Tests
+namespace Microsoft.DotNet.XHarness.Apple.Tests
 {
     public class AppInstallerTests : IDisposable
     {
