+++ tests/Microsoft.DotNet.XHarness.Apple.Tests/AppRunnerTests.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.DotNet.XHarness.Apple;
 using Microsoft.DotNet.XHarness.Common.Execution;
 using Microsoft.DotNet.XHarness.Common.Logging;
 using Microsoft.DotNet.XHarness.Common.Utilities;
@@ -20,7 +21,7 @@ using Microsoft.DotNet.XHarness.iOS.Shared.Utilities;
 using Moq;
 using Xunit;
 
-namespace Microsoft.DotNet.XHarness.iOS.Tests
+namespace Microsoft.DotNet.XHarness.Apple.Tests
 {
     public class AppRunnerTests : IDisposable
     {
